cross-macos: get aarch64-apple-darwin workspace check to 0 errors
Port the Darwin-specific platform surface so every #[cfg(target_os="macos")]
path compiles. Linux check remains green.

bun_sys::darwin
- __ulock_wait/__ulock_wait2/__ulock_wake + UL/ULOp packed flags (Futex)
- COPYFILE packed-struct flags, ftruncate/truncate re-exports
- kevent64/EVFILT/EV/NOTE were already present; add bun_sys::kevent()
  slice-wrapped Maybe form (sys.zig:2278) and bun_sys::c::kqueue()

bun_sys::c (translate-c surface)
- mach host/processor/vm primitives for os.cpus() and memory stats:
  host_processor_info, mach_host_self, mach_task_self, vm_deallocate,
  processor_cpu_load_info, natural_t, mach_msg_type_number_t,
  PROCESSOR_CPU_LOAD_INFO[_COUNT], HOST_VM_INFO64*, CPU_STATE_*
- sysctl/sysctlbyname, struct_loadavg, sockaddr_dl
- copyfile/clonefile/fcopyfile/fclonefileat + COPYFILE_* constants
- _dyld_image_count/_dyld_get_image_header/_dyld_get_image_vmaddr_slide
- memset_pattern{4,8,16}, _NSGetEnviron, truncate
- UTIME_NOW/UTIME_OMIT (libc crate omits these on Apple)

bun_sys::macho (new)
- mach_header_64, load_command, segment_command_64, LC_SEGMENT_64
- LoadCommandIterator port of std.macho for crash_handler ASLR-unslid addrs

bun_sys::posix
- sysctlbyname Maybe-wrapped (std.posix.sysctlbynameZ)
- timeval re-export
- s_islnk/s_isdir/s_isreg widen S_IF* to u32 (mode_t is u16 on Darwin)

bun_errno::darwin_errno
- harmonize SystemErrno API with linux_errno: MAX:u16, from_raw, init(i64),
  GetErrno trait + impls so bun_sys re-exports compile uniformly
- S::IS* predicates take u32 (callers normalize st_mode to bun_core::Mode)
- inline libuv synthetic errno values (ECHARSET/ENONET/EREMOTEIO/EUNATCH)

bun_io kqueue path
- demote ApplyAction const-generic enum to runtime param (adt_const_params
  is nightly-only and ACTION is never used in a type position)
- kevent64_s.ident is u64 on Darwin, fix cast
- Errno::EINTR (E-prefixed), Tag::kevent casing, Error::from_errno()

bun_threading::Futex (Darwin)
- wire to bun_sys::darwin::__ulock_*; replace transmute<i32,E> (size
  mismatch on #[repr(u16)]) with E::from_raw

bun_spawn_sys
- local extern decls for posix_spawn_file_actions_add{inherit,chdir}_np
  and POSIX_SPAWN_SETSID (libc crate omits the _np Darwin variants)
- posix_spawn argv/envp: cast *const *const c_char -> *const *mut c_char
  (Darwin libc header is non-const; never written)

callers
- mode_t is u16 on Darwin: cast at libc boundary in mkdir/chmod/umask/etc;
  add 'as u32' on st_mode at S::IS* call sites in node_fs/fetch/blob/shell
- libc crate flattens Darwin st_*timespec into st_*time/st_*time_nsec
  (PosixStat.rs, resolver/stat_hash.rs)
- DT_WHT not in libc crate on Apple; literal 14
- lchmod/fdatasync/getcontext: local unsafe extern decls (in libSystem,
  omitted by libc crate)
- pthread_threadid_np for thread_id() on Darwin (was unstable
  ThreadId::as_u64)
- image/codecs: BackendError::split() to reshape CoreGraphics
  Result<T,BackendError> into caller's Result<Option<T>,Error>
- EventLoopHandle::from_tag_ptr inverse for dns libinfo path
- cli/open: const-block promote &'static [&'static ZStr] arrays
- assorted: SocketAddress sa_family_t u8->u16 cast, HostProcess vm ptr
  cast + envp iter, blob clonefile borrowck reshape, ffi xcrun gated
  TODO until bun::spawn_sync lands

heap_breakdown.rs: drop local macro re-exports that collided with the fn
of the same name in the value namespace (only compiled on macOS).

+++ src/runtime/image/Image.rs
@@ -832,7 +832,7 @@ impl Image {
     pub fn clipboard_change_count(_: &JSGlobalObject, _: &CallFrame) -> JsResult<JSValue> {
         #[cfg(any(target_os = "macos", windows))]
         {
-            return Ok(JSValue::js_number(codecs::system_backend::clipboard_change_count()));
+            return Ok(JSValue::js_number(codecs::system_backend::clipboard_change_count() as f64));
         }
         #[cfg(not(any(target_os = "macos", windows)))]
         Ok(JSValue::js_number(-1.0))
